+++ src/lib.rs
@@ -1,155 +1,3 @@
-/// World is a trait that allows for plugins to define methods
-/// they expect from the instance (via super traits on impl) and
-/// to ensure at compile-time that every loaded plugin is OK with
-/// the World instance.
-/// 
-/// In other words, each plugin can request the World to have certain
-/// methods, if it needs to talk with another plugin via the World instance.
-/// 
-/// Example:
-/// ```rust
-/// struct Ping {
-///     pinged: bool
-/// }
-/// struct PingerPlugin;
-/// struct WaitingPlugin;
-/// trait WorldRequirements {
-///     fn get_ping_field(&self) -> &Ping;
-///     fn set_ping_field(&mut self, ping: &Ping);
-/// }
-/// 
-/// // Here, the compiler knows that PingerPlugin requires the World instance
-/// // to implement the WorldRequirements trait. So, if we'd attempt to add
-/// // PingerPlugin when the World instance doesn't implement this trait, 
-/// // we would obtain a compile-error.
-/// impl<SD: SharedData + WorldRequirements> Plugin<SD> for PingerPlugin {
-///     fn build() -> Self {
-///         Self
-///     }
-/// }
-/// 
-/// // Same thing for this plugin
-/// impl<SD: SharedData + WorldRequirements> Plugin<SD> for WaitingPlugin {
-///     fn build() -> Self {
-///         Self
-///     }
-/// }
-/// 
-/// // Note that this World instance doesn't implement the 
-/// // WorldRequirements trait!
-/// struct NonValidWorld;
-/// 
-/// impl SharedData for NonValidWorld {
-///     fn build() -> Self {
-///         Self
-///     }
-/// }
-/// 
-/// fn main() {
-///     let app = App::new_with_world::<NonValidWorld>();
-///     // Compile error! Plugins just can't be added into the plugin
-///     // typed system because type requirements aren't met!
-///     let app = app.add_plugins::<(PingerPlugin, WaitingPlugin)>();
-///     app.build().run();
-/// }
-/// ```
-pub trait SharedData {
-    fn build() -> Self;
-}
-
-pub trait Plugin<SD: SharedData> {
-    /// Return an instance of the Plugin
-    fn build() -> Self;
-    /// As soon as the App built instance is runned
-    /// (`app.run()`), this method will be executed.
-    /// Just after that all plugins finished this method,
-    /// the update loop begins. Note that `pre_update` method is
-    /// called after all plugins finish executing this method.
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    #[inline(always)]
-    fn startup(&mut self) {}
-    /// Schedule, this method runs each loop execution. It gets runned after
-    /// the execution of `update_exit_status_with_world` of previous loop execution.
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    #[inline(always)]
-    fn pre_update(&mut self) {}
-    /// Schedule, gets runned after the `pre_update` method.
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    #[inline(always)]
-    fn update(&mut self) {}
-    /// Schedule, gets runned after the `update` method, and is
-    /// followed by `access_ref_world`.
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    #[inline(always)]
-    fn post_update(&mut self) {}
-    /// Any plugin can access World instance by reference
-    /// once a frame. 
-    /// Plugins can for instance update their own data based on
-    /// the accessed data, or execute instant logic with it.
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    /// Schedule, runs after `post_update`.
-    #[inline(always)]
-    fn access_ref_world(_world: &SD) {}
-    /// Similar to the `access_ref_world` method, but provides
-    /// a mutable reference to the World instance. 
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    /// Schedule, runs after `access_ref_world`.
-    #[inline(always)]
-    fn access_mutref_world(_world: &mut SD) {}
-    /// This blank implementation will be optimized away by the compiler if
-    /// the plugin doesn't reimplement this method.
-    /// As soon as this method changes should_exit value to true, the ECS calls
-    /// `on_exit` methods just before exiting the loop.
-    /// 
-    /// Here's what's needed to change ShouldExit struct value to true:
-    /// ```rust
-    /// should_exit.request_exit();
-    /// ```
-    /// 
-    /// Note that as soon as exit is requested, it can't be cancelled in any way!
-    /// Schedule, runs after `access_mutref_world`.
-    #[inline(always)]
-    fn update_exit_status(&self, _should_exit: &mut ShouldExit) {}
-    /// Same as for the `update_exit_status` method, but the plugin can access World
-    /// too, so it can operate according to more informations.
-    /// Schedule, runs after `update_exit_status`.
-    #[inline(always)]
-    fn update_exit_status_with_world(&self, _should_exit: &mut ShouldExit, _world: &SD) {}
-    /// This method gets called as soon as the ECS holder is dropped.
-    /// There's no guaranties that all Plugins will have their exit
-    /// method called.
-    /// Note that the only two ways this function can get runned are:
-    /// - ECS holder drops
-    /// - a plugin has requested exit
-    /// This method runs after `update_exit_status_with_world` schedule if
-    /// a plugin has requested shutdown.
-    #[inline(always)]
-    fn on_exit(&mut self) {}
-}
-
-/// ECS holder field, that is distributed for all plugins
-/// that want to be able to gracefully shutdown the ECS.
-/// As soon as its value is true, all plugins have their
-/// `on_exit` method called, followed by loop breaking and
-/// then app shutdown.
-pub struct ShouldExit(bool);
-
-impl Default for ShouldExit {
-    fn default() -> Self {
-        Self(false)
-    }
-}
-
-impl ShouldExit {
-    /// Changes ShouldExit struct value to true, 
-    /// causing a graceful shutdown.
-    pub fn request_exit(&mut self) {
-        self.0 = true;
-    }
-}
+pub mod plugin;
+pub mod shared_data;
+pub mod app;
